Use a generator function to allow multiple async states
The previous version of async-state used deep-extend. The main reason
for adding deep-extend was to allow users to create their own namespaces
in case multiple modules in the same app used async-state. This was a
disaster when adding objects like `http.IncomingMessage` to the state,
as deep-extend isn't intended to clone such objects.

The commits instead adds a generator function to allow multiple modules
to each use their own async state inside the same app.